Version 2024.08.21 Queries: Apply refactoring Batches: Destroy refactoring Jobs: JobInfo added in all Execute methods Network: refactoring Editor graph: Sync label added Unity 2022 support added Use DestroyHierarchy instead of Destroy in addons

+++ Addons/Prototype/Units/Runtime/Aspects/UnitCommandGroupAspect.cs
@@ -19,8 +19,8 @@ namespace ME.BECS.Units {
         public readonly ref readonly ListAuto<Ent> readUnits => ref this.groupDataPtr.Read(this.ent.id, this.ent.gen).units;
         public readonly ref MemArrayAuto<Ent> targets => ref this.groupDataPtr.Get(this.ent.id, this.ent.gen).targets;
         public readonly ref readonly MemArrayAuto<Ent> readTargets => ref this.groupDataPtr.Read(this.ent.id, this.ent.gen).targets;
-        public readonly ref float volume => ref this.groupDataPtr.Get(this.ent.id, this.ent.gen).volume;
-        public readonly ref readonly float readVolume => ref this.groupDataPtr.Read(this.ent.id, this.ent.gen).volume;
+        public readonly ref uint volume => ref this.groupDataPtr.Get(this.ent.id, this.ent.gen).volume;
+        public readonly ref readonly uint readVolume => ref this.groupDataPtr.Read(this.ent.id, this.ent.gen).volume;
         public readonly ref Ent nextChainTarget => ref this.groupDataPtr.Get(this.ent.id, this.ent.gen).nextChainTarget;
         public readonly ref Ent prevChainTarget => ref this.groupDataPtr.Get(this.ent.id, this.ent.gen).prevChainTarget;
